+++ jupman_tools.py
@@ -319,7 +319,7 @@ def init(jupman):
     # Method as in https://github.com/spatialaudio/nbsphinx/issues/305#issuecomment-506748814
 
     def _from_notebook_node(self, nb, resources, **kwargs):
-        info('patched preprocessing method')            
+        info('patched nbsphinx from_notebook_node')
                                 
         for f in jupman.preprocessors:
             nb, resources = f.preprocess(nb, resources=resources)
@@ -343,7 +343,7 @@ class JupmanPreprocessor(Preprocessor):
     def preprocess(self, nb, resources):
         """ @since 3.2 """
 
-        info("****  Jupman: preprocessing notebook *****")
+        info("JupmanPreprocesser running...")
         
         """Careful path *includes* part of docname:
         {
@@ -702,10 +702,10 @@ class Jupman:
                 show = self.ipynb_show_hide
                 sol_class = 'jupman-sol-code'
 
-            s = """<div id="%s" class="jupman-sol %s" onclick="jupman.toggleSolution('%s');" >""" % (sid,sol_class,sid)
-            s += """\n    <input type="button" class="jupman-sol-label" """
-            s += """        data-jupman-show="%s" data-jupman-hide="%s" value="%s"/>""" % (show, hide, show)
-            s += """\n    <div class="jupman-sol-content" style="display:none">""" 
+            #s = """<div id="%s" class="jupman-sol %s">""" % (sid,sol_class)
+            s = """\n    <a class="jupman-sol jupman-sol-toggler" """
+            s += """\n      onclick="jupman.toggleSolution(this);" """
+            s += """        data-jupman-show="%s" data-jupman-hide="%s">%s</a>""" % (show, hide, show)            
             
             ret = new_raw_cell()
             ret.metadata.format = "text/html"
@@ -743,7 +743,7 @@ class Jupman:
                         nb.cells.append(before_cell(cell_counter, cell.cell_type))
                         cell.source = _cancel_tags(cell.source, self.tags)
                         nb.cells.append(cell)
-                        nb.cells.append(after_cell())                    
+                        #nb.cells.append(after_cell())                    
                     nb.cells.append(stripped_cell)
                 else:
                     nb.cells.append(cell)
@@ -752,7 +752,7 @@ class Jupman:
                     if website:
                         nb.cells.append(before_cell(cell_counter, cell.cell_type))
                         nb.cells.append(cell)
-                        nb.cells.append(after_cell()) 
+                        #nb.cells.append(after_cell()) 
                     else:
                          # substitues with newline, otherwise it shows 'Type markdown or latex'   
                         stripped_cell.source = re.sub( self.markdown_answer,